+++ CHANGELOG.md
@@ -5,6 +5,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](https://semver.org/).
 
+## [v0.2.3] - 2022-03-14
+
+### Added
+
+- feat: Export BoxedObjectStream so that users can implement Layer (#147)
+
+## [v0.2.2] - 2022-03-14
+
+### Fixed
+
+- services/fs: Refactor via tokio::fs (#142)
+- fix: Stat root should return a dir object (#143)
+
 ## [v0.2.1] - 2022-03-10
 
 ### Added
@@ -144,6 +157,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/).
 
 Hello, OpenDAL!
 
+[v0.2.3]: https://github.com/datafuselabs/opendal/compare/v0.2.2...v0.2.3
+[v0.2.2]: https://github.com/datafuselabs/opendal/compare/v0.2.1...v0.2.2
 [v0.2.1]: https://github.com/datafuselabs/opendal/compare/v0.2.0...v0.2.1
 [v0.2.0]: https://github.com/datafuselabs/opendal/compare/v0.1.4...v0.2.0
 [v0.1.4]: https://github.com/datafuselabs/opendal/compare/v0.1.3...v0.1.4